+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `group()` method for grouping results
+- `facet()` method for getting facet counts
 
 ## [1.7.2] - 2020-11-27
 ### Fixed